# backend/openapi.yaml
openapi: 3.0.3
info:
  title: HR / E-Learning API
  version: "1.0.0"
  description: |
    OpenAPI spec generated from your NestJS controllers & DTOs (User, Auth, Admin, Feedback, AuditLog, Backup).
    Sensitive fields (passwordHash, mfaSecret, mfaBackupCodes) are intentionally excluded from public response schemas.
servers:
  - url: http://localhost:3786
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          example: "Alice Example"
        email:
          type: string
          format: email
          example: "alice@example.com"
        password:
          type: string
          example: "P@ssw0rd!"
        role:
          type: string
          enum: [Student, Instructor, Admin]
          example: Student
        learningPreferences:
          type: array
          items:
            type: string
        subjectsOfInterest:
          type: array
          items:
            type: string
        expertise:
          type: array
          items:
            type: string
        profileImage:
          type: string
      required:
        - name
        - email
        - password
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Student, Instructor, Admin]
        profileImage:
          type: string
    PublicUserDto:
      type: object
      properties:
        _id:
          type: string
          example: "64b8f7a0c2f4e9b3f1a2d3c4"
        name:
          type: string
          example: "Alice Example"
        email:
          type: string
          format: email
          example: "alice@example.com"
        role:
          type: string
          example: "Student"
        isEmailVerified:
          type: boolean
          example: false
        unreadNotificationCount:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedUserDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicUserDto'
        total:
          type: integer
          example: 234
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 12
        limit:
          type: integer
          example: 20

    # -------------------------
    # Auth
    # -------------------------
...
    UpdateUserRoleBodyDto:
      type: object
      properties:
        role:
          type: string
          enum: [Student, Instructor, Admin]
      required: [role]
    AnnounceAllDto:
      type: object
      properties:
        message:
          type: string
      required: [message]
    AnnounceRoleDto:
      type: object
      properties:
        role:
          type: string
          enum: [Student, Instructor, Admin]
        message:
          type: string
      required: [role, message]
...
    SecurityOverviewDto:
      type: object
      properties:
        failedLogins:
          type: integer
          example: 120
        unauthorizedAttempts:
          type: integer
          example: 15
        distinctSuspiciousIps:
          type: integer
          example: 7
        windowFrom:
          type: string
          format: date-time
        windowTo:
          type: string
          format: date-time

    # -------------------------
    # Backup
    # -------------------------
...
  parameters:
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
paths:
...
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: User by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
...
  /admin/create-user:
    post:
      tags:
        - admin
      summary: Create a user as admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
  /admin/{id}:
    patch:
      tags:
        - admin
      summary: Update user by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
    delete:
      tags:
        - admin
      summary: Delete user by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
  /admin/{id}/soft-delete:
    post:
      tags:
        - admin
      summary: Soft-delete user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Soft-deleted
          content:
            application/json:
              schema:
                type: object
  /admin/{id}/restore:
    post:
      tags:
        - admin
      summary: Restore soft-deleted user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                type: object
  /admin/users:
    get:
      tags:
        - admin
      summary: List users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserDto'
  /admin/users/{id}/role:
    patch:
      tags:
        - admin
      summary: Update user role
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleBodyDto'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserDto'
  /admin/users/export:
    get:
      tags:
        - admin
      summary: Export users CSV (download)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV file response (binary)
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
  /admin/security:
    get:
      tags:
        - admin
      summary: Security overview
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Security overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityOverviewDto'
  /admin/announce/all:
    post:
      tags:
        - admin
      summary: Send announcement to all users
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnounceAllDto'
      responses:
        '200':
          description: Announcement queued
          content:
            application/json:
              schema:
                type: object
  /admin/announce/role:
    post:
      tags:
        - admin
      summary: Send announcement to a role
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnounceRoleDto'
      responses:
        '200':
          description: Announcement queued
          content:
            application/json:
              schema:
                type: object
  /admin/users/{id}/lock:
    patch:
      tags:
        - admin
      summary: Lock a user (prevent login)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Locked
          content:
            application/json:
              schema:
                type: object
  /admin/users/{id}/unlock:
    patch:
      tags:
        - admin
      summary: Unlock a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Unlocked
          content:
            application/json:
              schema:
                type: object

  # -------------------------
  # Backup
  # -------------------------
...